Fix parameter type for generateCode
