+++ frontend/src/generateCode.ts
@@ -1,23 +1,16 @@
 import toast from "react-hot-toast";
 import { WS_BACKEND_URL } from "./config";
 import { USER_CLOSE_WEB_SOCKET_CODE } from "./constants";
+import { FullGenerationSettings } from "./types";
 
 const ERROR_MESSAGE =
   "Error generating code. Check the Developer Console AND the backend logs for details. Feel free to open a Github issue.";
 
 const STOP_MESSAGE = "Code generation stopped";
 
-export interface CodeGenerationParams {
-  generationType: "create" | "update";
-  image: string;
-  resultImage?: string;
-  history?: string[];
-  // isImageGenerationEnabled: boolean; // TODO: Merge with Settings type in types.ts
-}
-
 export function generateCode(
   wsRef: React.MutableRefObject<WebSocket | null>,
-  params: CodeGenerationParams,
+  params: FullGenerationSettings,
   onChange: (chunk: string) => void,
   onSetCode: (code: string) => void,
   onStatusUpdate: (status: string) => void,
